+++ src/main/java/net/smokeybbq/bittermelon/medical/simulation/compartments/SimpleCompartment.java
@@ -4,11 +4,11 @@ import net.smokeybbq.bittermelon.medical.substance.Substance;
 
 public class SimpleCompartment extends Compartment {
 
-    public SimpleCompartment(String name, double volume) {
+    public SimpleCompartment(String name, float volume) {
         super(name, volume);
     }
 
-    public double getDerivative(double sourceConcentration, Substance drug) {
+    public float getDerivative(float sourceConcentration, Substance drug) {
         return bloodFlow * sourceConcentration - (bloodFlow / 2) * getConcentration(drug);
     }
 }
